# contacts-api/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Contacts API",
    "version": "1.0.0",
    "description": "API for managing contacts"
  },
  "paths": {
    "/contacts": {
      "get": {
        "summary": "Retrieve all contacts",
        "responses": {
          "200": {
            "description": "A list of contacts",
            "content": {
              "application/json": {
                "example": [
                  {
                    "_id": "61d9d9d9d9d9d9d9d9d9d9d9",
                    "firstName": "John",
                    "lastName": "Doe",
                    "email": "john.doe@example.com",
                    "favoriteColor": "Blue",
                    "birthday": "2000-01-01"
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new contact",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "email": { "type": "string" },
                  "favoriteColor": { "type": "string" },
                  "birthday": { "type": "string", "format": "date" }
                },
                "required": ["firstName", "lastName", "email", "favoriteColor", "birthday"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created successfully"
          }
        }
      }
    },
    "/contacts/{id}": {
      "put": {
        "summary": "Update an existing contact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Contact updated successfully" }
        }
      },
      "delete": {
        "summary": "Delete a contact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Contact deleted successfully" }
        }
      }
    }
  }
}